% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schedule}
\alias{schedule}
\title{Semester schedule}
\format{
This is a [data.frame] with 69 rows and 35 variables:
\describe{
 \item{Term}{The year and semester code.}
 \item{Code}{The course id number including the academic program abbreviation and the course number (e.g., 'BIOL 101')}
 \item{Title}{The officieal course title}
 \item{Credits}{The number of credit hours }
 \item{Instructor}{The last name of the individual tasked with teaching the course.}
 \item{Enrollment}{The number of students in the course.}
 \item{From}{Start date (as date object).}
 \item{To}{End date (as date object).}
 \item{Days}{Weekdays as (MWF, TR, etc.)}
 \item{Start}{Time to start the class each session (e.g., 1200).}
 \item{End}{Time the course ends each time (e.g., 1300).}
 \item{Location}{The building and room number for the course.}
}
}
\usage{
schedule
}
\description{
An example of a set of course offerings for a semester including the
  names of the courses, their titles, instructors, and meeting locations.
}
\keyword{datasets}
