% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcu_courses}
\alias{vcu_courses}
\title{VCU Course Bulletin}
\format{
this is a [data.frame] with 2863 rows and 6 variables:
\describe{
  \item{code}{The internal designation for the class (e.g., 'BIOL 101').}
  \item{title}{The official course title.}
  \item{credits}{The number of credits that the university gives for this course.}
  \item{bulletin}{The raw university bulletin text for the class}
  \item{contact}{Contact hours per week as lecture, workship, laboratory, etc.}
  \item{prerequs}{Any pre-requisites required for this course.}
}
}
\usage{
vcu_courses
}
\description{
This R data set has listings for 2863 course bulletins with relevant
  information for each class that is on the books.  These classes were
  downloaded from the VCU Bulletin in spring 2025 and the classes
  were parsed from the raw PDF.
}
\keyword{datasets}
