% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_select_n_credits.R
\name{weights_select_n_credits}
\alias{weights_select_n_credits}
\title{Parse choose N credits}
\usage{
weights_select_n_credits(
  x,
  courses,
  max_enum = 15,
  trials = 1e+05,
  seed = NULL
)
}
\arguments{
\item{x}{The mapping string to be parsed.}

\item{courses}{A [data.frame] that has columns named `code` for the course code and `credtis` for the number of credits for the class (required).}

\item{max_enum}{The number of courses where the estimator will go from brute force combinatorial approch to Monte Carlo approximations.
This gives 2^k-1 subsets and for something like 15 is 32,767, which is reasonable for R on a desktop.  Beyond that, it does the approximation.}

\item{trials}{The number of iterations to run for Monte Carlo approximation}

\item{seed}{A random starting seed}
}
\value{
A [data.frame] where each row is a course with columns for `Course` and `Weight` as defined by the rules above.
}
\description{
This helper function steps in when there is a "choose N credits" from the list
 I had to update it a bit to handle cases where the list of courses is large
 (VT & GMU I'm looking at you and your funky class options...) to use a
 Monte Carlo approximation when N or the number of credtis gets too big
}
\examples{
data( vcu_courses )
encoding <- "CHEM101[6cred]CHEM102[6cred]PHYS101[6cred]MATH201[6cred]ENGL201"
weights_select_n_credits(encoding, vcu_courses)
weights_select_n_credits(encoding, vcu_courses, max_enum=3)
}
