% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_weight.R
\name{course_weight}
\alias{course_weight}
\title{Parses Course For Relative Weighing}
\usage{
course_weight(x, courses, code = "Course", credits = "Credits")
}
\arguments{
\item{x}{The mapping string to be parsed.}

\item{courses}{A [data.frame] that has course code and credtis}

\item{code}{The name of the column in [courses] where course codes are found (default: Course)}

\item{credits}{The name of the column in [courses] with the number of credits for the course (default: Credits)}
}
\value{
A [data.frame] where each row is a course with columns for `Weight` and `Course` as defined by the rules above.
}
\description{
This function takes a course listing and parses it based
  on the number of credits that the linguistic weighings
  will be base upon.  The credit weighing is used to scale
  the linguistic mappings.
}
\details{
There are several ways to encode the requirements for courses.  These
  include:
  \describe{
  \item{ABCD101}{The single course gets the full credit weight.}
  \item{ABCD101|ABCD201}{Both courses get half of their credit weights (OR)}
  \item{ABCD101&ABCD210}{Each course is weighed as full credit but must be together {AND}}
  \item{ABCD101[2]ABCD102[2]ABCD103[2]ABCD104[2] }{Pick two from the list of courses and they get full weight but then do all combinations and weigh all of them in proportion to their frequency.}
  \item{ABCD101[6c]ABCD102[6c]ABCD103[6c]ABCD104[6c]}{Pick a minimum of 6 credits from the list and get credit equivallents and then do all combinations and weight portitional to their frequency.}
  }
}
